fix: using zustando for states and use client for nextJs
