+++ src/components/Toast/Toast.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React from 'react';
 import { CheckCircle, WarningCircle, Info, X } from 'phosphor-react';
