--- src/components/Toast/ToastProvider.test.tsx
@@ -1,187 +1,0 @@
-import { render, screen, act, fireEvent } from '@testing-library/react';
-import '@testing-library/jest-dom';
-import { ToastProvider, useToast } from './ToastProvider';
-
-beforeEach(() => {
-  /* eslint-disable-next-line no-undef */
-  Object.defineProperty(global, 'crypto', {
-    value: {
-      randomUUID: () => 'mock-uuid-123-1-1',
-    },
-    writable: true,
-  });
-});
-
-afterEach(() => {
-  jest.restoreAllMocks();
-});
-
-describe('ToastProvider', () => {
-  it('should render children', () => {
-    render(
-      <ToastProvider>
-        <div data-testid="child">Test Child</div>
-      </ToastProvider>
-    );
-
-    expect(screen.getByTestId('child')).toBeInTheDocument();
-  });
-
-  it('should add and remove toasts', () => {
-    const TestComponent = () => {
-      const { addToast, removeToast } = useToast();
-
-      return (
-        <div>
-          <button
-            onClick={() => addToast({ title: 'Test Toast' })}
-            data-testid="add-toast"
-          >
-            Add Toast
-          </button>
-          <button
-            onClick={() => removeToast('mock-uuid-123-1-1')}
-            data-testid="remove-toast"
-          >
-            Remove Toast
-          </button>
-        </div>
-      );
-    };
-
-    render(
-      <ToastProvider>
-        <TestComponent />
-      </ToastProvider>
-    );
-
-    act(() => {
-      fireEvent.click(screen.getByTestId('add-toast'));
-    });
-
-    expect(screen.getByText('Test Toast')).toBeInTheDocument();
-
-    // Remove the toast
-    act(() => {
-      fireEvent.click(screen.getByTestId('remove-toast'));
-    });
-
-    expect(screen.queryByText('Test Toast')).not.toBeInTheDocument();
-  });
-
-  it('should render toast with all props', () => {
-    const TestComponent = () => {
-      const { addToast } = useToast();
-
-      return (
-        <button
-          onClick={() =>
-            addToast({
-              title: 'Full Toast',
-              description: 'This is a description',
-              variant: 'outlined',
-              action: 'success',
-              position: 'top-right',
-            })
-          }
-          data-testid="add-full-toast"
-        >
-          Add Full Toast
-        </button>
-      );
-    };
-
-    render(
-      <ToastProvider>
-        <TestComponent />
-      </ToastProvider>
-    );
-
-    act(() => {
-      fireEvent.click(screen.getByTestId('add-full-toast'));
-    });
-
-    expect(screen.getByText('Full Toast')).toBeInTheDocument();
-    expect(screen.getByText('This is a description')).toBeInTheDocument();
-  });
-});
-
-describe('useToast', () => {
-  it('should throw error when used outside of ToastProvider', () => {
-    const originalError = console.error;
-    console.error = jest.fn();
-
-    const TestComponent = () => {
-      useToast();
-      return null;
-    };
-
-    expect(() => render(<TestComponent />)).toThrow(
-      'useToast must be used within a ToastProvider'
-    );
-
-    console.error = originalError;
-  });
-
-  it('should provide addToast and removeToast functions', () => {
-    const TestComponent = () => {
-      const { addToast, removeToast } = useToast();
-      return (
-        <div>
-          <button onClick={() => addToast({ title: 'Test' })} data-testid="add">
-            Add
-          </button>
-          <button onClick={() => removeToast('123')} data-testid="remove">
-            Remove
-          </button>
-        </div>
-      );
-    };
-
-    render(
-      <ToastProvider>
-        <TestComponent />
-      </ToastProvider>
-    );
-
-    expect(screen.getByTestId('add')).toBeInTheDocument();
-    expect(screen.getByTestId('remove')).toBeInTheDocument();
-  });
-
-  it('should call removeToast when Toast onClose is triggered', () => {
-    const TestComponent = () => {
-      const { addToast } = useToast();
-
-      return (
-        <button
-          onClick={() => addToast({ title: 'Toast to close' })}
-          data-testid="add-toast"
-        >
-          Add Toast
-        </button>
-      );
-    };
-
-    render(
-      <ToastProvider>
-        <TestComponent />
-      </ToastProvider>
-    );
-
-    act(() => {
-      fireEvent.click(screen.getByTestId('add-toast'));
-    });
-
-    const toastElement = screen.getByText('Toast to close');
-    expect(toastElement).toBeInTheDocument();
-
-    const closeButton = screen.getByRole('button', {
-      name: /dismiss notification/i,
-    });
-    act(() => {
-      fireEvent.click(closeButton);
-    });
-
-    expect(screen.queryByText('Toast to close')).not.toBeInTheDocument();
-  });
-});
