--- src/components/Toast/ToastProvider.tsx
@@ -1,74 +1,0 @@
-import { createContext, useContext, useState, ReactNode, useMemo } from 'react';
-import { Toast } from './Toast';
-
-type ToastPosition =
-  | 'top-left'
-  | 'top-center'
-  | 'top-right'
-  | 'bottom-left'
-  | 'bottom-center'
-  | 'bottom-right'
-  | 'default';
-
-type ToastData = {
-  id: string;
-  title: string;
-  description?: string;
-  variant?: 'solid' | 'outlined';
-  action?: 'warning' | 'success' | 'info';
-  position?: ToastPosition;
-};
-
-type ToastContextType = {
-  addToast: (toast: Omit<ToastData, 'id'>) => void;
-  removeToast: (id: string) => void;
-};
-
-const ToastContext = createContext<ToastContextType | undefined>(undefined);
-
-export const ToastProvider = ({ children }: { children: ReactNode }) => {
-  const [toasts, setToasts] = useState<ToastData[]>([]);
-
-  const addToast = (toast: Omit<ToastData, 'id'>) => {
-    const id = crypto.randomUUID();
-    setToasts((prev) => [...prev, { id, ...toast }]);
-  };
-
-  const removeToast = (id: string) => {
-    setToasts((prev) => prev.filter((t) => t.id !== id));
-  };
-
-  const contextValue = useMemo(
-    () => ({
-      addToast,
-      removeToast,
-    }),
-    [addToast, removeToast]
-  );
-
-  return (
-    <ToastContext.Provider value={contextValue}>
-      {children}
-
-      {toasts.map((toast) => (
-        <Toast
-          key={toast.id}
-          title={toast.title}
-          description={toast.description}
-          variant={toast.variant}
-          action={toast.action}
-          position={toast.position}
-          onClose={() => removeToast(toast.id)}
-        />
-      ))}
-    </ToastContext.Provider>
-  );
-};
-
-export const useToast = (): ToastContextType => {
-  const context = useContext(ToastContext);
-  if (!context) {
-    throw new Error('useToast must be used within a ToastProvider');
-  }
-  return context;
-};
