+++ src/components/Toast/utils/ToastStore.ts
@@ -1,0 +1,40 @@
+import { create } from 'zustand';
+
+type ToastPosition =
+  | 'top-left'
+  | 'top-center'
+  | 'top-right'
+  | 'bottom-left'
+  | 'bottom-center'
+  | 'bottom-right'
+  | 'default';
+
+export type ToastData = {
+  id: string;
+  title: string;
+  description?: string;
+  variant?: 'solid' | 'outlined';
+  action?: 'warning' | 'success' | 'info';
+  position?: ToastPosition;
+};
+
+type ToastStore = {
+  toasts: ToastData[];
+  addToast: (toast: Omit<ToastData, 'id'>) => void;
+  removeToast: (id: string) => void;
+};
+
+export const useToastStore = create<ToastStore>((set) => ({
+  toasts: [],
+  addToast: (toast) => {
+    const id = crypto.randomUUID();
+    set((state) => ({
+      toasts: [...state.toasts, { id, ...toast }],
+    }));
+  },
+  removeToast: (id) => {
+    set((state) => ({
+      toasts: state.toasts.filter((t) => t.id !== id),
+    }));
+  },
+}));
