+++ src/components/Toast/utils/Toaster.tsx
@@ -1,0 +1,32 @@
+'use client';
+
+import { useToastStore } from './ToastStore';
+import { Toast } from '../Toast';
+
+export const Toaster = () => {
+  const toasts = useToastStore((state) => state.toasts);
+  const removeToast = useToastStore((state) => state.removeToast);
+
+  return (
+    <>
+      {toasts.map((toast) => (
+        <Toast
+          key={toast.id}
+          title={toast.title}
+          description={toast.description}
+          variant={toast.variant}
+          action={toast.action}
+          position={toast.position}
+          onClose={() => removeToast(toast.id)}
+        />
+      ))}
+    </>
+  );
+};
+
+export const useToast = () => {
+  const addToast = useToastStore((state) => state.addToast);
+  const removeToast = useToastStore((state) => state.removeToast);
+
+  return { addToast, removeToast };
+};
