+++ src/index.ts
@@ -4,5 +4,5 @@ export * from './components/IconRoundedButton/IconRoundedButton';
 export * from './components/SelectionButton/SelectionButton';
 export * from './components/Text/Text';
 export * from './components/Toast/Toast';
-export * from './components/Toast/ToastProvider';
+export * from './components/Toast/utils/Toaster';
 export * from './components/Table/Table';
